Add validation on whether car was found.
With the current implementation, if were doing a post to the cars endpoint with a new id, oldCar becomes undefined and index returns a value of -1. This causes a -1 being passed into the splice function which actually starts the splice from the last index of the array. What happens here is a replacement of the last index with the new entry.

The code I'm proposing does a validation check on the oldCar, and in the case the oldCar is undefined, we will just push the car into the array. This clears up my confusion as to why the last index was being replaced rather than an actual "create" happening on the post request when supplying an id.